GH #42: Showing human errors when EA is not found.

+++ scripts/run_backtest.sh
@@ -121,7 +121,7 @@ while getopts $ARGS arg; do
     e) # EA name.
       EA_NAME=${OPTARG}
       EA_PATH="$(find "$ROOT" '(' -name "*$EA_NAME*.ex4" -o -name "*$EA_NAME*.ex5" ')' -print -quit)"
-      [ -f "$EA_PATH" ]
+      [ -f "$EA_PATH" ] || { echo "Error: EA file ($EA_NAME) not found in '$ROOT'!" >&2; exit 1; }
       cp $VFLAG "$EA_PATH" "$TERMINAL_DIR/MQL4/Experts";
       ini_set "^TestExpert" "$(basename "${EA_PATH%.*}")" "$TESTER_INI"
       ;;
